+++ packages/ai/src/ui-message-stream/handle-ui-message-stream-finish.ts
@@ -5,10 +5,7 @@ import {
 } from '../ui/process-ui-message-stream';
 import { UIMessage } from '../ui/ui-messages';
 import { ErrorHandler } from '../util/error-handler';
-import {
-  InferUIMessageStreamPart,
-  UIMessageStreamPart,
-} from './ui-message-stream-parts';
+import { InferUIMessageChunk, UIMessageChunk } from './ui-message-chunks';
 
 export function handleUIMessageStreamFinish<UI_MESSAGE extends UIMessage>({
   messageId,
@@ -17,9 +14,13 @@ export function handleUIMessageStreamFinish<UI_MESSAGE extends UIMessage>({
   onError,
   stream,
 }: {
-  stream: ReadableStream<InferUIMessageStreamPart<UI_MESSAGE>>;
+  stream: ReadableStream<InferUIMessageChunk<UI_MESSAGE>>;
 
-  messageId: string;
+  /**
+   * The message ID to use for the response message.
+   * If not provided, no id will be set for the response message.
+   */
+  messageId?: string;
 
   /**
    * The original messages.
@@ -46,18 +47,47 @@ export function handleUIMessageStreamFinish<UI_MESSAGE extends UIMessage>({
      */
     responseMessage: UI_MESSAGE;
   }) => void;
-}): ReadableStream<InferUIMessageStreamPart<UI_MESSAGE>> {
-  if (onFinish == null) {
-    return stream;
+}): ReadableStream<InferUIMessageChunk<UI_MESSAGE>> {
+  // last message is only relevant for assistant messages
+  let lastMessage: UI_MESSAGE | undefined =
+    originalMessages?.[originalMessages.length - 1];
+  if (lastMessage?.role !== 'assistant') {
+    lastMessage = undefined;
+  } else {
+    // appending to the last message, so we need to use the same id
+    messageId = lastMessage.id;
   }
 
-  const lastMessage = originalMessages?.[originalMessages.length - 1];
+  const idInjectedStream = stream.pipeThrough(
+    new TransformStream<
+      InferUIMessageChunk<UI_MESSAGE>,
+      InferUIMessageChunk<UI_MESSAGE>
+    >({
+      transform(chunk, controller) {
+        // when there is no messageId in the start chunk,
+        // but the user checked for persistence,
+        // inject the messageId into the chunk
+        if (chunk.type === 'start') {
+          const startChunk = chunk as UIMessageChunk & { type: 'start' };
+          if (startChunk.messageId == null && messageId != null) {
+            startChunk.messageId = messageId;
+          }
+        }
+
+        controller.enqueue(chunk);
+      },
+    }),
+  );
+
+  if (onFinish == null) {
+    return idInjectedStream;
+  }
 
   const state = createStreamingUIMessageState<UI_MESSAGE>({
     lastMessage: lastMessage
       ? (structuredClone(lastMessage) as UI_MESSAGE)
       : undefined,
-    messageId, // will be overridden by the stream
+    messageId: messageId ?? '', // will be overridden by the stream
   });
 
   const runUpdateMessageJob = async (
@@ -70,32 +100,13 @@ export function handleUIMessageStreamFinish<UI_MESSAGE extends UIMessage>({
   };
 
   return processUIMessageStream<UI_MESSAGE>({
-    stream: stream.pipeThrough(
-      new TransformStream<
-        InferUIMessageStreamPart<UI_MESSAGE>,
-        InferUIMessageStreamPart<UI_MESSAGE>
-      >({
-        transform(chunk, controller) {
-          // when there is no messageId in the start chunk,
-          // but the user checked for persistence,
-          // inject the messageId into the chunk
-          if (chunk.type === 'start') {
-            const startChunk = chunk as UIMessageStreamPart & { type: 'start' };
-            if (startChunk.messageId == null) {
-              startChunk.messageId = messageId;
-            }
-          }
-
-          controller.enqueue(chunk);
-        },
-      }),
-    ),
+    stream: idInjectedStream,
     runUpdateMessageJob,
     onError,
   }).pipeThrough(
     new TransformStream<
-      InferUIMessageStreamPart<UI_MESSAGE>,
-      InferUIMessageStreamPart<UI_MESSAGE>
+      InferUIMessageChunk<UI_MESSAGE>,
+      InferUIMessageChunk<UI_MESSAGE>
     >({
       transform(chunk, controller) {
         controller.enqueue(chunk);