+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/xai
 
+## 2.0.0-beta.5
+
+### Patch Changes
+
+- 66b9661: feat (provider/xai): export XaiProviderOptions
+- Updated dependencies [1b101e1]
+  - @ai-sdk/openai-compatible@1.0.0-beta.4
+
+## 2.0.0-beta.4
+
+### Patch Changes
+
+- 9e986f7: feat (provider/xai): add grok-4 model id
+
 ## 2.0.0-beta.3
 
 ### Patch Changes